requirements.txt
torch >= 2.1.0
hydra-core >= 1.3.2
transformer_lens >= 1.12.1
pytest >= 7.4.4 # Required by transformer_lens